+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
 ## [0.7.16] 2020-02-03
 
 ### Added